Improve performance by disabling CanvasRenderer GameObjects
Canvas rendering was occurring despite attempts to disable canvas component and to set alpha on CanvasRenderer Components. CanvasRenderer Components also can't be disabled, as they are not Behaviors, so the current solution is use SelectedGameObjectCuller to disable the entire canvas gameobject in the station prefabs based on the selection status of their ancestor MetroGame component.