bug fixes and performance tweaks
Train.prefab: removed mesh collider from train geometry object and replaced with box collider on train prefab root.  This saves performance and gets rid of exception from the MRTK interactable script

MetroManager.cs: Games get created with metro manager as their transform parent to reduce clutter from 100 games in the scene hierarchy

Track.cs: Call UpdateControlPoints() to populate cp array when beginning to track the first track segment of a transport line  this was causing a bunch of errors when creating new lines.

TrackSegment.cs:  Doesn't create the mesh for interaction if there is no line to interact with.  This might not be the best way to implement this but it would cause 300 errors on game start, 3 per game.