Added moving trains between lines
Clicking on a train selects it in its gameInstance, clicking on a track segment with a train selected will have the train transfer tracks after unloading all of its passengers at the next station it reaches.  A ghost train is added following the left hand while a train is selected to indicate the selection, I plan to make the ghost train appear on the track currently being hovered over to further indicate this.  Currently there is no way to deselect a train, after it is clicked it must be moved, so this will also need to be implemented